+++ android/app/src/main/java/com/ismin/projectapp/Town.kt
@@ -2,4 +2,4 @@ package com.ismin.projectapp
 
 import java.io.Serializable
 
-data class Town (val city: String, val population: Double, val country: String) : Serializable
+data class Town(val city: String, val population: Int, val country: String) : Serializable
